The branch was 649 commits behind and conflicted in 34 files, every one an
eager ModelData site that master had also edited. Each took master's
version with this PR's single-line change re-applied:
ModelData = <serialize call>  ->  ModelDataProvider = () => <same expr>.

Also:
- FasterWhisper built its metadata in a one-line initializer, so its
  ModelData = SerializeForMetadata() was missed by both this PR and its
  guard. Converted, and the guard regex now also matches an assignment
  after a separator (",", "{", "("); it fails with the old line restored.
- The branch had stripped the UTF-8 BOM from 70 files it touched. Master's
  BOM is restored, so the diff against master is only the ModelData change.

Verified: net10.0 test build succeeds; EagerModelDataAssignmentGuardTests,
ModelMetadataLazyModelData and the affected model families: 103 passed.

+++ src/Evaluation/Statistics/DeLongTest.cs
@@ -142,7 +142,7 @@ public class DeLongTest<T> : IStatisticalTest<T>
                 else if (Math.Abs(posProb - negProb) < 1e-10) sum += 0.5;
             }
         }
-        return sum / (posIndices.Count * negIndices.Count);
+        return sum / ((double)posIndices.Count * negIndices.Count);
     }
 
     private double[] ComputeV10(T[] predictions, List<int> posIndices, List<int> negIndices)
